Improve error messages
Currently when exception is thrown it is not handled
and as a result the stack trace is printed when the programs
exits due to the exception. Although quite useful the stack trace
may look scary to non-experienced developers or even may give the
impression that the application is poorly written.

Modify the exception handling so all exception are cough and their
message is printed to the console.